[refactor] Main menu screen to have more consistent naming
