+++ App/Navigation/NavigationRouter.js
@@ -21,7 +21,7 @@ import APITestingScreen from '../Containers/APITestingScreen'
 import ThemeScreen from '../Containers/ThemeScreen'
 import DeviceInfoScreen from '../Containers/DeviceInfoScreen'
 
-import BarMenuScreen from '../Containers/BarMenuScreen'
+import MenuBarScreen from '../Containers/MenuBarScreen'
 import MenuBeersScreen from '../Containers/MenuBeersScreen'
 import MenuShotsScreen from '../Containers/MenuShotsScreen'
 import MenuCocktailsScreen from '../Containers/MenuCocktailsScreen'
@@ -48,7 +48,7 @@ class NavigationRouter extends Component {
             <Scene key='apiTesting' component={APITestingScreen} title='API Testing' />
             <Scene key='theme' component={ThemeScreen} title='Theme' />
 
-            <Scene key='barMenu' component={BarMenuScreen} title='Bar Menu' />
+            <Scene key='barMenu' component={MenuBarScreen} title='Bar Menu' />
             <Scene key='beersMenu' component={MenuBeersScreen} title='Beers Menu' />
             <Scene key='shotsMenu' component={MenuShotsScreen} title='Shots Menu' />
             <Scene key='cocktailsMenu' component={MenuCocktailsScreen} title='Cocktails Menu' />
